+++ src/domain/infrastructure/twitch/twitch.ts
@@ -103,3 +103,25 @@ export interface GetGame {
 }
 
 export type GetGamesResponse = SearchPaginationResponse<GetGame>;
+
+//-----------------------------------------------------------------------------
+//--------------------------- Get Games -----------------------------------
+//-----------------------------------------------------------------------------
+
+export interface GetFollowedStreams {
+    id: string;
+    user_id: string;
+    user_login: string;
+    user_name: string;
+    game_id: string;
+    game_name: string;
+    type: string;
+    title: string;
+    viewer_count: number;
+    started_at: string;
+    language: string;
+    thumbnail_url: string;
+    tag_ids: string[];
+}
+
+export type GetFollowedStreamsResponse = SearchPaginationResponse<GetFollowedStreams>;