+++ src/infrastructure/twitch/twitchRepository.ts
@@ -1,6 +1,8 @@
 import { AxiosInstance, AxiosResponse } from 'axios';
-import { CLIENT_ID } from '../../config';
+import { API_BASE_URL, CLIENT_ID, OAUTH_BASE_URL } from '../../config';
 import {
+    GetFollowedStreams,
+    GetFollowedStreamsResponse,
     GetGame,
     GetGamesResponse,
     GetStream,
@@ -15,9 +17,6 @@ import { createAxiosInstance, getToken } from './twitchHelpers';
 
 let apiInstance: AxiosInstance;
 let oAuthInstance: AxiosInstance;
-
-export const OAUTH_BASE_URL = 'https://id.twitch.tv/oauth2';
-const API_BASE_URL = 'https://api.twitch.tv/helix';
 
 export const MAX_INTEGER_VALUE = 100;
 
@@ -38,7 +37,9 @@ const getApiInstance = (): AxiosInstance => {
 export const revokeToken = async (): Promise<void> => {
     try {
         const token: string = await getToken();
-        await getOAuthInstance().post(`${OAUTH_BASE_URL}/revoke?client_id=${CLIENT_ID}&token=${token}`);
+        await getOAuthInstance().post(
+            `${OAUTH_BASE_URL}/revoke?client_id=${CLIENT_ID}&token=${token}`,
+        );
     } catch (e) {
         console.error('Error revoking token', e?.response?.data || e.message);
         throw e;
@@ -50,6 +51,7 @@ export const validateToken = async (): Promise<ValidateTokenResponse> => {
         const response: AxiosResponse<ValidateTokenResponse> = await getOAuthInstance().get(
             `${OAUTH_BASE_URL}/validate`,
         );
+
         return response.data;
     } catch (e) {
         console.error('Error validating the token', e?.response?.data || e.message);
@@ -57,7 +59,11 @@ export const validateToken = async (): Promise<ValidateTokenResponse> => {
     }
 };
 
-export const getUserFollows = async (fromId = '', toId = '', first = 20): Promise<GetUserFollow[]> => {
+export const getUserFollows = async (
+    fromId = '',
+    toId = '',
+    first = 20,
+): Promise<GetUserFollow[]> => {
     if (!fromId && !toId) {
         throw new Error('At minimum, from_id or to_id must be provided for a query to be valid.');
     }
@@ -106,7 +112,8 @@ export const getStreams = async (
             usersIds.length && usersIds.map((userId) => url.searchParams.append('user_id', userId));
             gamesIds.length && gamesIds.map((gameId) => url.searchParams.append('game_id', gameId));
             languages.length && languages.map((lang) => url.searchParams.append('language', lang));
-            usersLogins.length && usersLogins.map((userLogin) => url.searchParams.append('user_login', userLogin));
+            usersLogins.length &&
+                usersLogins.map((userLogin) => url.searchParams.append('user_login', userLogin));
             url.searchParams.append('first', first.toString());
             cursor && url.searchParams.append('after', cursor);
 
@@ -125,7 +132,8 @@ export const getStreams = async (
 
 export const getUsers = async (ids: string[] = [], login: string[] = []): Promise<GetUser[]> => {
     try {
-        const response: AxiosResponse<GetUsersResponse> = await getApiInstance().get(`${API_BASE_URL}/users?
+        const response: AxiosResponse<GetUsersResponse> = await getApiInstance()
+            .get(`${API_BASE_URL}/users?
             ${ids.map((id) => `&id=${id}`).join('')}
             ${login.map((login) => `&login=${login}`).join('')}`);
 
@@ -154,6 +162,27 @@ export const getGames = async (ids: string[] = [], names: string[] = []): Promis
         return games;
     } catch (e) {
         console.error('Error getting games', e?.response?.data || e.message);
+        throw e;
+    }
+};
+
+export const getFollowedStreams = async (userId: string): Promise<GetFollowedStreams[]> => {
+    try {
+        let streams: GetFollowedStreams[] = [];
+        let response: AxiosResponse<GetFollowedStreamsResponse>;
+        let cursor: string | undefined;
+        do {
+            response = await getApiInstance()
+                .get(`${API_BASE_URL}/streams/followed?user_id=${userId}
+            ${cursor ? `&after=${cursor}` : ''}`);
+
+            cursor = response.data?.pagination?.cursor;
+            streams = [...streams, ...response.data.data];
+        } while (response.data?.pagination?.cursor);
+
+        return streams;
+    } catch (e) {
+        console.error('Error getting followed streams', e?.response?.data || e.message);
         throw e;
     }
 };